removes disused componenets / organises pages into pages folder
