+++ frontend/src/Pages/SingleGroupPage.js
@@ -1,24 +1,20 @@
-import VertNavbar from '../VertNavBar/VertNavBar';
-import { useParams } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
-import isTokenValid from '../Utility/isTokenValid';
-import getSessionUserID from '../Utility/getSignedInUser_id';
-import '../../Pages/style.css'
-import BlackboardHeader from '../blackboardHeader/blackboardHeader';
-import { useLocation } from 'react-router-dom';
-import calculateGroupStats from './groupLeaderBoard';
+import { useParams, useLocation } from 'react-router-dom';
+import VertNavbar from '../components/VertNavBar/VertNavBar';
+import isTokenValid from '../components/Utility/isTokenValid';
+import getSessionUserID from '../components/Utility/getSignedInUser_id';
+import BlackboardHeader from '../components/blackboardHeader/blackboardHeader';
+import calculateGroupStats from '../components/groupLeaderboard/groupLeaderBoard';
+import '../Pages/style.css'
 
 const SingleGroupPage = ({ navigate }) => {
 const { pubGroupId } = useParams();
 const [pubGroupData, setPubGroupData] = useState(null);
 const [token, setToken] = useState(window.localStorage.getItem("token"));
-const [isLoggedIn, setIsLoggedIn] = useState(isTokenValid(token));
+const [isLoggedIn] = useState(isTokenValid(token));
 
 const [groupWagers,setGroupWagers] = useState([]);
-const [groupMembers,setGroupMembers] = useState([]);
 const [groupLeaderboardData,setGroupLeaderboardData] = useState([])
-
-
 
 const location = useLocation();
 const expandedState = location.state?.expandedState;
@@ -138,89 +134,70 @@ const toggleGroupMembership = async () => {
 return (
 
 <div className='shade'>
-<div className={`page-content ${expanded ? 'shifted-content' : ''}`}>
-<div className='blackboard'>
-<VertNavbar expanded={expanded} toggleExpand={toggleExpand} />
-<BlackboardHeader expandedState={expanded}/> 
+  <div className={`page-content ${expanded ? 'shifted-content' : ''}`}>
+    <div className='blackboard'>
+      <VertNavbar expanded={expanded} toggleExpand={toggleExpand} />
+      <BlackboardHeader expandedState={expanded}/> 
 
-<div className='chalktitle'>{pubGroupData?.name}</div>
-    <div id='ButtonDiv'>
+      <div className='chalktitle'>{pubGroupData?.name}</div>
+      <div id='ButtonDiv'>
         <button onClick={toggleGroupMembership} className='orange_Button'> {!isGroupMember ? 'Join':'Leave'} </button>
-    </div>
-    <div style={{display:'flex'}}>
-    <div style={{width:'50%'}}>
-        <div id='TopStats'>
+      </div>
+      <div style={{display:'flex'}}>
+        <div style={{width:'50%'}}>
+          <div id='TopStats'>
             <span className="orange-chalk">Biggest Boozer</span> <br/>
             <div className='small-chalk' style={{textAlign:'center',marginBottom:'0px'}}>
-              {mostTotalWinsUsername}</div>
+              {mostTotalWinsUsername}
+            </div>
             <span className="orange-chalk">Biggest Weeper</span> <br/>
             <div className='small-chalk' style={{textAlign:'center',marginBottom:'0px'}}> {mostTotallossesUsername} </div>
             <span className="orange-chalk">Free Pints won in {pubGroupData?.name}</span> <br/>
             <div className='small-chalk' style={{textAlign:'center',marginBottom:'0px'}}> {totalPints}</div>
-        </div>
+          </div>
 
-        <div id='UserBets'>
+          <div id='UserBets'>
             <div className="orange-chalk">Latest Wagers...</div>
             {/* Displays most recent 10 wagers in group */}
             {sortedWagersData.slice(0,10)?.map((wager) => (
             <div key={wager._id}>
-            {wager.winner ? (wager.winner._id===wager.peopleInvolved[0]._id ? 
-            // Winner & Person 0 is the Winner
-            <div className='small-chalk'>{wager.peopleInvolved[0].username} won their wager with {wager.peopleInvolved[1].username} that {wager.description}</div> :
-            // Winner & Person 1 is the winner
-            <div className='small-chalk'>{wager.peopleInvolved[1].username} won their wager with {wager.peopleInvolved[0].username} that {wager.description}</div>)
-            : 
-            // No Winner            
-            <div className='small-chalk'>
-            {wager.peopleInvolved[0].username} bet {wager.peopleInvolved[1].username} that {wager.description}
+              {wager.winner ? (wager.winner._id===wager.peopleInvolved[0]._id ? 
+              // Winner & Person 0 is the Winner
+              <div className='small-chalk'>{wager.peopleInvolved[0].username} won their wager with {wager.peopleInvolved[1].username} that {wager.description}</div> :
+              // Winner & Person 1 is the winner
+              <div className='small-chalk'>{wager.peopleInvolved[1].username} won their wager with {wager.peopleInvolved[0].username} that {wager.description}</div>)
+              : 
+              // No Winner            
+                <div className='small-chalk'>
+                  {wager.peopleInvolved[0].username} bet {wager.peopleInvolved[1].username} that {wager.description}
+                </div>
+              }
             </div>
-            }</div>
-
-))}
-
-        </div>
-    </div>
-    <div className='post-it' style={{width:'50%'}}>
-    <p className="note" style={{width:'90%'}}>
-
-    <span className="penfont-large centered">Leaderboard</span>
-    {OnlyOneMember || AllMembersOnZeroPercent ? (
-        <div style={{textAlign:'center'}} className="penfont-small">
-          Resolve a wager with other members to climb the Leaderboard
-        </div>
-) : (
-      sortedWinPercent.map((item, index) => (
-          <div key={item.id}>
-              <span className="penfont-small">{index + 1} {item.username} - {item.winPercentage}%</span>
+            ))}
           </div>
-    ))
-)}
-    
+        </div>
 
-      
+        <div className='post-it' style={{width:'50%'}}>
+          <p className="note" style={{width:'90%'}}>
+            <span className="penfont-large centered">Leaderboard</span>
+            {OnlyOneMember || AllMembersOnZeroPercent ? (
+            <div style={{textAlign:'center'}} className="penfont-small">
+              Resolve a wager with other members to climb the Leaderboard
+            </div>
+            ) : (
+            sortedWinPercent.map((item, index) => (
+            <div key={item.id}>
+              <span className="penfont-small">{index + 1} {item.username} - {item.winPercentage}%</span>
+            </div>
+            ))
+            )}
 
-
-
-    </p>
+          </p>
+        </div>
+      </div>
     </div>
+  </div>
 </div>
-
-
-
-
-  
-
-
-
-
-</div>
-</div>
-</div>
-
-
-
-
-);
-};
+)};
     
     export default SingleGroupPage;
