--- frontend/src/components/SearchBar/SearchBar.css
@@ -1,58 +1,0 @@
-
-.search-bar {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 20px;
-  }
-  
-  .search-container {
-    display: flex;
-    align-items: center;
-  }
-  
-  .search-field {
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #cd561b;
-    border-radius: 4px;
-    margin-right: 8px;
-  }
-  
-  .result-container {
-    margin-top: 10px;
-  }
-  
-  ul {
-    list-style: none;
-    padding: 0;
-  }
-  
-  .searchRes {
-    display: block;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    margin-bottom: 8px;
-    text-decoration: none;
-    color: #333;
-  }
-  
-  .searchRes:hover {
-    background-color: #eee9e7;
-  }
-  .searchMessage {
-    background-color: #fcfcfc;
-    color: #111111;
-    border: 1px solid #fefefe;
-    padding: 10px;
-    border-radius: 4px;
-    display: flex;
-    align-items: center;
-  }
-  
-  .close-icon {
-    margin-left: 8px;
-    cursor: pointer;
-  }
-
