--- frontend/src/components/header/header.css
@@ -1,31 +1,0 @@
-
-.HeaderMainBar {
-    color: #fff;
-    width: 100%;
-    text-align: right;
-    align-items: center; /* Center content vertically */
-    justify-content: flex-end; /* Align content to the right */
-  }
-  
-  .HeaderMainBar img {
-    position: fixed;
-    top: 0px;
-    right: 0px;
-    padding-top: 2px;
-    height: auto;
-    width: 160px;
-    padding-top: 10px;
-    padding-bottom: 0px;
-    padding-right: 50px;
-  }
-
-  .bottom-right-image {
-    position: fixed;
-    bottom: 0px; /* Adjust this value to position the image vertically */
-    right: 0px; /* Adjust this value to position the image horizontally */
-    z-index: -1; 
-    height: 350px;
-    width: auto;
-    opacity: 0.35;
-  }
-
