+++ frontend/src/components/leaderboard/leaderboard.js
@@ -1,37 +1,84 @@
-import React, { useEffect, useState } from 'react';
-import isTokenValid from '../Utility/isTokenValid';
-import VertNavbar from '../VertNavBar/VertNavBar';
-import BlackboardHeader from '../blackboardHeader/blackboardHeader';
-import Leaderboard from '../stats/leaderboard';
-import '../../Pages/style.css';
-import { useLocation } from 'react-router-dom';
+import React, { useState, useEffect } from 'react';
+import calculateUserStats from '../stats/calculateuserstats';
+import './leaderstats.css'
 
-
-const LeaderBoardPage = ({ navigate }) => {
-  const [token, setToken] = useState(window.localStorage.getItem("token"));
-  const [wagers, setWagers] = useState([]);
-  const [isLoggedIn, setIsLoggedIn] = useState(isTokenValid(token));
-  const location = useLocation();
-  const expandedState = location.state?.expandedState;
-  const [expanded, setExpanded] = useState(expandedState !== undefined ? expandedState : true);
-
-  const toggleExpand = () => { setExpanded(!expanded); };
+const Leaderboard = () => {
+  const [userToken, setUserToken] = useState(window.localStorage.getItem("token"));
+  const [pintsData, setpintsData] = useState([]);
+  const [userObject, setuserObject] = useState([]);
+  const [userWithMostLosses, setUserWithMostLosses] = useState(null);
+  const [userWithMostWins, setUserWithMostWins] = useState(null);
 
   useEffect(() => {
-    if (!isLoggedIn) { navigate('/', { state: { expandedState: expanded } }); }
-  }, [navigate, isLoggedIn]);
+    const fetchPints = async () => {
+      try {
+        const response = await fetch(`/pints`, {
+        headers: { Authorization: `Bearer ${userToken}` }
+        });
+        if (!response.ok) {throw new Error('Network response was not ok');}
+        const fetchedData = await response.json();
+        setpintsData(fetchedData.pints);
+      } catch (error) {console.error('Error fetching pint data:', error);}
+    };
+
+    fetchPints(); 
+  }, [userToken]); 
+
+  useEffect(() => {
+    const userObjectArray = calculateUserStats(pintsData)
+      .sort((a, b) => b.winPercentage - a.winPercentage)
+      .slice(0, 10)
+    setuserObject(userObjectArray);
+
+    // Find user with the most pintsOwedNotClaimed
+    const userWithMostLosses = userObjectArray.reduce((prev, current) =>
+      prev.betsLost > current.betsLost ? prev : current, {});
+
+    setUserWithMostLosses(userWithMostLosses);
+
+    // Find user with the highest pintsOwnedClaimed
+    const userWithMostWins = userObjectArray.reduce((prev, current) =>
+      prev.betsWon > current.betsWon ? prev : current, {});
+
+    setUserWithMostWins(userWithMostWins);
+
+  }, [pintsData]);
 
   return (
-    <div className='shade'>
-      <div className={`page-content ${expanded ? 'shifted-content' : ''}`}>
-        <div className='blackboard'>
-          <VertNavbar expanded={expanded} toggleExpand={toggleExpand} />
-          <BlackboardHeader expandedState={expanded}/>
-          <Leaderboard />
-        </div>
+    <div>
+      <div className='chalktitle'>Leaderboard</div>
+      <span className='chalktitle'>Diamond Geezers...</span>
+      <ul>
+        {userObject.map((user, index) => (
+          <li key={index} className='leadertext'>
+            {index + 1} {user.username} - {user.winPercentage}% Win rate
+          </li>
+        ))}
+      </ul>
+
+      <div>
+      <span className='chalktitle'>Top Boozer</span>
+        {userWithMostWins && (
+          <p className='leadertext'>
+            {userWithMostWins.username} has won {userWithMostWins.betsWon} pints <br/>
+            <span className='leadercomments'>(Absolutely Outrageous)</span>
+          </p>
+        )}
       </div>
+
+      <div>
+      <span className='chalktitle'>Top Weeper</span>
+        {userWithMostLosses && (
+          <p className='leadertext'>
+            <b>{userWithMostLosses.username}</b> has had to buy {userWithMostLosses.betsLost} pints <br/>
+            <span className='leadercomments'>(What an absolute mug!!)</span>
+          </p>
+        )}
+      </div>
+
     </div>
+
   );
 };
 
-export default LeaderBoardPage;
+export default Leaderboard;
