+++ frontend/src/components/myAccountPage/UnresolvedWagers.js
@@ -1,9 +1,8 @@
 import React, { useEffect, useState } from 'react';
-// import React, { useEffect, useState } from 'react';
-import "../MyAccountPage.css"
-import './notification.css'
-import getSessionUserID from '../../Utility/getSignedInUser_id';
 import { Link } from 'react-router-dom';
+import getSessionUserID from '../Utility/getSignedInUser_id';
+import { FaCircle } from "react-icons/fa";
+import '../myAccountPage/MyAccountPage.css'
 
 const UnresolvedWagers = ({ navigate, wagers, expandedState }) => {
 	const [token, setToken] = useState(window.localStorage.getItem("token"));
@@ -21,7 +20,7 @@ const UnresolvedWagers = ({ navigate, wagers, expandedState }) => {
 				
 					{wagers.map((wager) => (
 					<div key={wager._id}>
-						<Link to ={`/wager/${wager._id}`} className='individualwagerlink' state = {{expandedState: expanded }}>Time's Up! Who won the wager that {wager.description}? </Link>
+						<Link to ={`/wager/${wager._id}`} className='individualwagerlink' state = {{expandedState: expanded }}><FaCircle  size={10}/> Time's Up! Who won the wager that {wager.description}? </Link>
 					</div>))}
 				</div>
