--- frontend/src/components/myAccountPage/myAccountPageComponents/notification.css
@@ -1,28 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Caveat&display=swap');
-
-.notification{
-    font-size: 24px;
-    margin-bottom: 20px;
-    color: white;
-    font-family: 'Permanent Marker';
-} 
-
-.notificationdetails {
-    font-size: 1.25rem;
-    margin-bottom: 20px;
-    color: white;
-    text-decoration: none;
-    font-family: 'Permanent Marker';
-    padding-left: 40px;
-}
-
-.notificationdetails a,.notificationdetails:visited{
-    color: white;
-    text-decoration: none;
-}
-
-
-.individualwagerlink:hover {
-    text-decoration: underline;
-  }
-
