+++ frontend/src/components/singleWagerPageComponents/SingleOngoingWager.js
@@ -1,4 +1,4 @@
-import getSessionUserID from "../../Utility/getSignedInUser_id";
+import getSessionUserID from "../Utility/getSignedInUser_id";
 import React, { navigate, useEffect, useState } from 'react';
 import { useNavigate } from "react-router-dom";
