+++ frontend/src/components/singleWagerPageComponents/SinglePendingWager.js
