+++ frontend/src/components/singleWagerPageComponents/SingleWagerRequest.js
