+++ frontend/src/components/singleWagerPageComponents/WagerDetails.js
@@ -1,4 +1,4 @@
-import getSessionUserID from "../../Utility/getSignedInUser_id";
+import getSessionUserID from "../Utility/getSignedInUser_id";
 import React, { navigate } from 'react';
 import { useNavigate } from "react-router-dom";
