+++ frontend/src/components/stats/calculateuserstats.js
