--- frontend/src/components/wagers/NewWagerForm.css
@@ -1,18 +1,0 @@
-.NewWager {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 50vh;
-  }
-  
-
-  .NewWager form {
-    width: 260px; 
-    text-align: center;
-  }
-  
-  .NewWager input {
-    margin-bottom: 10px;
-  }
-
